chore: 기존 Header.jsx 컴포넌트 삭제 후 ResponsiveHeader.jsx 를 Header.jsx로 변경하여 사용, 개발툴 관련 코드 및 경로/파일 제거(react buddy)
