--- src/dev/index.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import { useInitial } from './useInitial';
-
-const ComponentPreviews = React.lazy(() => import('./previews'));
-
-export { ComponentPreviews, useInitial };
