+++ src/main.jsx
@@ -3,17 +3,13 @@ import ReactDOM from 'react-dom/client';
 import App from './App.jsx';
 import './index.css';
 import { BrowserRouter } from 'react-router-dom';
-import { DevSupport } from '@react-buddy/ide-toolbox';
-import { ComponentPreviews, useInitial } from './dev/index.js';
 import { NextUIProvider } from '@nextui-org/react';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <NextUIProvider>
       <BrowserRouter>
-        <DevSupport ComponentPreviews={ComponentPreviews} useInitialHook={useInitial}>
-          <App />
-        </DevSupport>
+        <App />
       </BrowserRouter>
     </NextUIProvider>
   </React.StrictMode>,
